+ add new graph to plot the latency histogram accounting for transactions that were cache hits * change the refreshCounter in Channel to use an array of already-created refresh transactions rather than an array of counters + add more power figures to the power stream + add unprocessed data to the power stream to allow calculations at post-processing time + attempt to use the JOBID from the torque server to number the output files, if available * check on output streams individually to make sure that they are created and opened properly * fix LRU copy constructor to properly relink the blocks + add boost::pool<> as the memory manager for Transactions

+++ src/Statistics.cpp
@@ -282,7 +282,7 @@ ostream &DRAMsimII::operator<<(ostream &os, const Statistics &statsLog)
 	os << statsLog.readCount << " " << statsLog.writeCount << " " << statsLog.readCount + statsLog.writeCount << endl;
 #endif
 
-	Statistics::WeightedAverage<unsigned> averageLatency;
+	Statistics::WeightedAverage<double> averageLatency;
 	os << "----Transaction Latency";
 	for (unordered_map<unsigned, unsigned>::const_iterator currentValue = statsLog.transactionExecution.begin(); currentValue != statsLog.transactionExecution.end(); currentValue++)
 	{
